hackathon-II/hackathon-todo-phase4/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.32
psycopg2-binary>=2.9.10
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.6.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.14.0
pytest==8.3.4
httpx==0.28.0
python-dotenv==1.0.1
mcp==1.26.0
openai==1.99.9
openai-agents==0.2.7